+++ src/targets/gpu/gemm_impl.cpp
@@ -592,16 +592,15 @@ static void gemm_save_solution(context& ctx,
                                const std::vector<shape>& input_shapes,
                                int32_t solution_idx)
 {
-    ctx.get_problem_cache().insert(
-        "rocblas", gemm_problem(output_shape, input_shapes), solution_idx);
+    ctx.problem_cache_insert("rocblas", gemm_problem(output_shape, input_shapes), solution_idx);
 }
 #endif
 
-int32_t gemm_default_solution(context& ctx,
+int32_t gemm_default_solution(const context& ctx,
                               const shape& output_shape,
                               const std::vector<shape>& input_shapes)
 {
-    auto sol = ctx.get_problem_cache().get("rocblas", gemm_problem(output_shape, input_shapes));
+    auto sol = ctx.problem_cache_get("rocblas", gemm_problem(output_shape, input_shapes));
     if(sol.has_value())
         return sol->to<int32_t>();
     return 0;