+++ src/targets/gpu/include/migraphx/gpu/gemm_impl.hpp
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2015-2025 Advanced Micro Devices, Inc. All rights reserved.
+ * Copyright (c) 2015-2026 Advanced Micro Devices, Inc. All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -88,7 +88,7 @@ int32_t gemm_finalize(context& ctx,
                       bool compute_fp32,
                       int32_t solution_idx);
 
-int32_t gemm_default_solution(context& ctx,
+int32_t gemm_default_solution(const context& ctx,
                               const shape& output_shape,
                               const std::vector<shape>& input_shapes);